+++ client/scripts/main.js
@@ -11,7 +11,7 @@ import { initialiseActions } from './lib/initialise';
 import { loadRam, saveRam, } from './lib/ramState';
 import { initialise as initDmdExport, save as saveFile } from './lib/pin2DmdExport';
 import { AudioOutput } from './lib/sound';
-import * as gamelist from './db/gamelist';
+import * as gamelist from './db';
 import { populateControlUiView, updateUiSwitchState } from './ui/control-ui';
 import * as emuDebugUi from './ui/oblivion-ui';
 
@@ -31,15 +31,13 @@ let missedDraw = 0;
 let rafId;
 
 function initialiseEmu(gameEntry) {
-  window.wpcInterface = {
-    romSelection,
-  };
-
   return document.fonts.load('24pt "Space Mono"')
     .catch((error) => {
       console.error('FONT_LOAD_FAILED', error);
     })
     .then(() => {
+      const selectElementRoot = document.getElementById('wpc-release-info');
+      selectElementRoot.innerHTML = 'WPC-Emu v' + global.RELEASE_VERSION;
       emuDebugUi.initialise();
       emuDebugUi.loadFeedback(gameEntry.name);
       return downloadFileFromUrlAsUInt8Array(gameEntry.rom.u06);
@@ -51,13 +49,8 @@ function initialiseEmu(gameEntry) {
       };
       return wpcEmuWebWorkerApi.initialiseEmulator(romData, gameEntry);
     })
-    .then(() => {
-      console.log('Successfully initialized emulator');
-      return wpcEmuWebWorkerApi.getVersion();
-    })
     .then((emuVersion) => {
-      const selectElementRoot = document.getElementById('wpc-release-info');
-      selectElementRoot.innerHTML = 'WPC-Emu v' + emuVersion;
+      console.log('Successfully initialized emulator', emuVersion);
       soundInstance = AudioOutput(gameEntry.audio);
       //NOTE: IIKS we pollute globals here
       window.wpcInterface = {